requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Django==3.2.9
django-environ==0.8.1
django-localflavor==3.1
django-phonenumber-field==6.0.0
gunicorn==20.1.0
icalendar==4.0.9
phonenumberslite==8.12.38